Model changes to reflect new design
