+++ models/ItemVariation.js
@@ -1,0 +1,22 @@
+const mongoose = require('mongoose');
+const Schema = mongoose.Schema;
+
+const ItemVariationSchema = new Schema({
+    name: {
+        type: String,
+        required: true
+    },
+    gender: {
+        type: [String]
+    },
+    size: {
+        type: [String]
+    },
+    typeColor: {
+        type: [String]
+    }
+});
+
+const ItemVariation = mongoose.model('ItemVariation', ItemVariationSchema);
+
+module.exports = ItemVariation;
